+++ src/StorageBase/Contract/IdentityDirectionKeyInterface.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Heptacom\HeptaConnect\Storage\Base\Contract;
+
+use Heptacom\HeptaConnect\Portal\Base\StorageKey\Contract\StorageKeyInterface;
+
+/**
+ * Identifies a directional identity.
+ */
+interface IdentityDirectionKeyInterface extends StorageKeyInterface
+{
+}